# *: optimize SortExec