--- a/distsql/xeval/eval_control_funcs.go
+++ b/distsql/xeval/eval_control_funcs.go
@@ -101,7 +101,7 @@ func (e *Evaluator) evalNullIf(expr *tipb.Expr) (d types.Datum, err error) {
 	if left.IsNull() || right.IsNull() {
 		return left, nil
 	}
-	x, err := left.CompareDatum(e.StatementCtx, right)
+	x, err := left.CompareDatum(e.StatementCtx, &right)
 	if err != nil {
 		return d, errors.Trace(err)
 	}