--- a/distsql/xeval/eval_control_funcs_test.go
+++ b/distsql/xeval/eval_control_funcs_test.go
@@ -79,7 +79,7 @@ func (s *testEvalSuite) TestEvalCaseWhen(c *C) {
 		result, err := xevaluator.Eval(tt.expr)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(xevaluator.StatementCtx, tt.result)
+		cmp, err := result.CompareDatum(xevaluator.StatementCtx, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}
@@ -136,7 +136,7 @@ func (s *testEvalSuite) TestEvalIf(c *C) {
 		result, err := xevaluator.Eval(tt.expr)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(xevaluator.StatementCtx, tt.result)
+		cmp, err := result.CompareDatum(xevaluator.StatementCtx, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}
@@ -176,7 +176,7 @@ func (s *testEvalSuite) TestEvalNullIf(c *C) {
 		result, err := xevaluator.Eval(tt.expr)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(xevaluator.StatementCtx, tt.result)
+		cmp, err := result.CompareDatum(xevaluator.StatementCtx, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}
@@ -211,7 +211,7 @@ func (s *testEvalSuite) TestEvalIfNull(c *C) {
 		result, err := xevaluator.Eval(tt.expr)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(xevaluator.StatementCtx, tt.result)
+		cmp, err := result.CompareDatum(xevaluator.StatementCtx, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}