--- a/distsql/xeval/eval_other_funcs_test.go
+++ b/distsql/xeval/eval_other_funcs_test.go
@@ -48,7 +48,7 @@ func (s *testEvalSuite) TestEvalCoalesce(c *C) {
 		result, err := xevaluator.Eval(tt.expr)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(xevaluator.StatementCtx, tt.result)
+		cmp, err := result.CompareDatum(xevaluator.StatementCtx, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}