--- a/executor/sort.go
+++ b/executor/sort.go
@@ -15,7 +15,9 @@ package executor
 
 import (
 	"container/heap"
+	"fmt"
 	"sort"
+	"time"
 
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/expression"
@@ -25,7 +27,7 @@ import (
 
 // orderByRow binds a row to its order values, so it can be sorted.
 type orderByRow struct {
-	key []types.Datum
+	key []*types.Datum
 	row Row
 }
 
@@ -105,17 +107,20 @@ func (e *SortExec) Next() (Row, error) {
 			}
 			orderRow := &orderByRow{
 				row: srcRow,
-				key: make([]types.Datum, len(e.ByItems)),
+				key: make([]*types.Datum, len(e.ByItems)),
 			}
 			for i, byItem := range e.ByItems {
-				orderRow.key[i], err = byItem.Expr.Eval(srcRow)
+				key, err := byItem.Expr.Eval(srcRow)
 				if err != nil {
 					return nil, errors.Trace(err)
 				}
+				orderRow.key[i] = &key
 			}
 			e.Rows = append(e.Rows, orderRow)
 		}
+		start := time.Now()
 		sort.Sort(e)
+		fmt.Printf("sort takes: %vns\n", time.Since(start).Nanoseconds())
 		e.fetched = true
 	}
 	if e.err != nil {
@@ -204,13 +209,14 @@ func (e *TopNExec) Next() (Row, error) {
 			// build orderRow from srcRow.
 			orderRow := &orderByRow{
 				row: srcRow,
-				key: make([]types.Datum, len(e.ByItems)),
+				key: make([]*types.Datum, len(e.ByItems)),
 			}
 			for i, byItem := range e.ByItems {
-				orderRow.key[i], err = byItem.Expr.Eval(srcRow)
+				key, err := byItem.Expr.Eval(srcRow)
 				if err != nil {
 					return nil, errors.Trace(err)
 				}
+				orderRow.key[i] = &key
 			}
 			if e.totalCount == e.heapSize {
 				// An equivalent of Push and Pop. We don't use the standard Push and Pop