--- a/expression/builtin_compare.go
+++ b/expression/builtin_compare.go
@@ -871,7 +871,7 @@ func refineConstantArg(con *Constant, op opcode.Op, ctx context.Context) *Consta
 		return con
 	}
 	datumInt := types.NewIntDatum(i64)
-	c, err := datumInt.CompareDatum(sc, con.Value)
+	c, err := datumInt.CompareDatum(sc, &con.Value)
 	if err != nil {
 		return con
 	}