--- a/expression/distsql_builtin_test.go
+++ b/expression/distsql_builtin_test.go
@@ -280,7 +280,7 @@ func (s *testEvalSuite) TestEval(c *C) {
 		result, err := expr.Eval(row)
 		c.Assert(err, IsNil)
 		c.Assert(result.Kind(), Equals, tt.result.Kind())
-		cmp, err := result.CompareDatum(sc, tt.result)
+		cmp, err := result.CompareDatum(sc, &tt.result)
 		c.Assert(err, IsNil)
 		c.Assert(cmp, Equals, 0)
 	}
@@ -445,7 +445,7 @@ func (s *testEvalSuite) TestEvalIsNull(c *C) {
 		result, err := expr.Eval(nil)
 		c.Assert(err, IsNil, Commentf("%v", tt))
 		c.Assert(result.Kind(), Equals, tt.result.Kind(), Commentf("%v", tt))
-		cmp, err := result.CompareDatum(sc, tt.result)
+		cmp, err := result.CompareDatum(sc, &tt.result)
 		c.Assert(err, IsNil, Commentf("%v", tt))
 		c.Assert(cmp, Equals, 0, Commentf("%v", tt))
 	}