--- a/store/localstore/local_region.go
+++ b/store/localstore/local_region.go
@@ -74,7 +74,7 @@ func (t *topnSorter) Less(i, j int) bool {
 		v1 := t.rows[i].key[index]
 		v2 := t.rows[j].key[index]
 
-		ret, err := v1.CompareDatum(t.ctx.sc, v2)
+		ret, err := v1.CompareDatum(t.ctx.sc, &v2)
 		if err != nil {
 			t.err = errors.Trace(err)
 			return true
@@ -123,7 +123,7 @@ func (t *topnHeap) Less(i, j int) bool {
 		v1 := t.rows[i].key[index]
 		v2 := t.rows[j].key[index]
 
-		ret, err := v1.CompareDatum(t.ctx.sc, v2)
+		ret, err := v1.CompareDatum(t.ctx.sc, &v2)
 		if err != nil {
 			t.err = errors.Trace(err)
 			return true