--- a/store/tikv/mock-tikv/topn.go
+++ b/store/tikv/mock-tikv/topn.go
@@ -49,7 +49,7 @@ func (t *topNSorter) Less(i, j int) bool {
 		v1 := t.rows[i].key[index]
 		v2 := t.rows[j].key[index]
 
-		ret, err := v1.CompareDatum(t.sc, v2)
+		ret, err := v1.CompareDatum(t.sc, &v2)
 		if err != nil {
 			t.err = errors.Trace(err)
 			return true
@@ -98,7 +98,7 @@ func (t *topNHeap) Less(i, j int) bool {
 		v1 := t.rows[i].key[index]
 		v2 := t.rows[j].key[index]
 
-		ret, err := v1.CompareDatum(t.sc, v2)
+		ret, err := v1.CompareDatum(t.sc, &v2)
 		if err != nil {
 			t.err = errors.Trace(err)
 			return true