--- a/util/types/compare_test.go
+++ b/util/types/compare_test.go
@@ -154,7 +154,7 @@ func compareForTest(a, b interface{}) (int, error) {
 	sc.IgnoreTruncate = true
 	aDatum := NewDatum(a)
 	bDatum := NewDatum(b)
-	return aDatum.CompareDatum(sc, bDatum)
+	return aDatum.CompareDatum(sc, &bDatum)
 }
 
 func (s *testCompareSuite) TestCompareDatum(c *C) {
@@ -177,11 +177,11 @@ func (s *testCompareSuite) TestCompareDatum(c *C) {
 	sc.IgnoreTruncate = true
 	for i, t := range cmpTbl {
 		comment := Commentf("%d %v %v", i, t.lhs, t.rhs)
-		ret, err := t.lhs.CompareDatum(sc, t.rhs)
+		ret, err := t.lhs.CompareDatum(sc, &t.rhs)
 		c.Assert(err, IsNil)
 		c.Assert(ret, Equals, t.ret, comment)
 
-		ret, err = t.rhs.CompareDatum(sc, t.lhs)
+		ret, err = t.rhs.CompareDatum(sc, &t.lhs)
 		c.Assert(err, IsNil)
 		c.Assert(ret, Equals, -t.ret, comment)
 	}