--- a/util/types/range.go
+++ b/util/types/range.go
@@ -133,7 +133,7 @@ func (ir *IndexRange) IsPoint(sc *variable.StatementContext) bool {
 		if a.Kind() == KindMinNotNull || b.Kind() == KindMaxValue {
 			return false
 		}
-		cmp, err := a.CompareDatum(sc, b)
+		cmp, err := a.CompareDatum(sc, &b)
 		if err != nil {
 			return false
 		}
@@ -202,7 +202,7 @@ func (ir *IndexRange) Align(numColumns int) {
 func (ir *IndexRange) PrefixEqualLen(sc *variable.StatementContext) (int, error) {
 	// Here, len(ir.LowVal) always equal to len(ir.HighVal)
 	for i := 0; i < len(ir.LowVal); i++ {
-		cmp, err := ir.LowVal[i].CompareDatum(sc, ir.HighVal[i])
+		cmp, err := ir.LowVal[i].CompareDatum(sc, &ir.HighVal[i])
 		if err != nil {
 			return 0, errors.Trace(err)
 		}