1. for struct `orderByRow` defined in file **executor/sort.go**, change `key []types.Datum` to `key []*types.Datum`
2. for function `Datum.CompareDatum()` defined in **util/types/datum.go**, change the function signature from `CompareDatum(sc *variable.StatementContext, ad Datum)` to `CompareDatum(sc *variable.StatementContext, ad *Datum)`

| num records | old sorter | new sorter | efficiency improved |
|:-------------|----------:|------------:|:--------------------|
| 10 | 1769ns | 1662ns | 6.44% |
| 100 | 28884ns | 22490ns | 28.43% |
| 1000 | 295691ns | 225859ns | 30.92% |
| 10000 | 3424773ns |2700828ns | 26.80% |